+++ src/gimmehttp/react/index.ts
@@ -1,0 +1,5 @@
+import GimmeHttp from './gimmehttp'
+export type { GimmeHttpProps, GimmeHttpRef } from './gimmehttp'
+
+export { GimmeHttp }
+export default GimmeHttp